Fixed time savings output of multi-threaded setup/processing pipeline (OVERLAP=ON) to no longer include wait time entering the map copy critical section.
This wait does not exist in the serial pipeline. It only exists in the multi-threaded pipeline and is a consequence of finishing setup for the concurrently running N number of threads - which is already accounted for. As such, it led to overcounting by as much as a factor of N*setup time.

The fix is to sum the setup times and the map copy times individually for each thread and not across the map copy critical section.